fix: pass 8 as "startWith" second parameter
I forgot to pass 8 as the second parameter of startWith method when qualify nodeName string. Because, it lead to proper context analysis. 8 is the length of "HyperMD-" string. It's written as the start name of each line that is treated as "block element", such us list, heading, and so on.

That string ("HyperMD-") is followed by one of block types. From that point, if you want to get the block type with startWith method (by immediately typing its block type like "list"), you have to pass the offset number, which is that length above. Otherwise, you get nothing but false return, not because of no type either, but because you are facing "HyperMD-" string.